CHROMIUM: mm: kstaled: a kswapd alternative
kstaled differs from kswapd in the following respects:

1) kstaled walks page tables directly to clear the accessed bit on
PTE. This allows kstaled to track the accesses of all mapped pages
with even less CPU overhead comparing with the rmap.

2) kstaled stores a N-bit age, i.e., timestamp of the last access,
for each page and sorts pages based on the age. This is a closer
approximation to the LRU, which reduces swap refaults.

3) kstaled scans at a fixed interval to prevent thrashing. If it
can't produce enough cold pages for reclaim to consume, allocation
will be throttled to allow the tab discarder to take action.

BUG=b:123039911
TEST=Ran platform.MemoryPressure and platform.MemoryStress
Test results can be found at go/cros-mm.

Change-Id: I608bc60d4a8679d2912568628ce144fe2d2adfef
Signed-off-by: Yu Zhao <yuzhao@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1789103
Reviewed-by: Yu Zhao <yuzhao@chromium.org>
Commit-Queue: Yu Zhao <yuzhao@chromium.org>
Tested-by: Yu Zhao <yuzhao@chromium.org>