UPSTREAM: drm/nouveau: Stop unsetting mstc->port, use malloc refs
Same as we did for i915, but for nouveau this time. Additionally, we
grab a malloc reference to the port that lasts for the entire lifetime
of nv50_mstc, which gives us the guarantee that mstc->port will always
point to valid memory for as long as the mstc stays around.

Signed-off-by: Lyude Paul <lyude@redhat.com>
Reviewed-by: Ben Skeggs <bskeggs@redhat.com>
Cc: Daniel Vetter <daniel@ffwll.ch>
Cc: David Airlie <airlied@redhat.com>
Cc: Jerry Zuo <Jerry.Zuo@amd.com>
Cc: Harry Wentland <harry.wentland@amd.com>
Cc: Juston Li <juston.li@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20190111005343.17443-16-lyude@redhat.com
(cherry picked from commit d79a3c52f34b648c56794e14364cc5169af2202f)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:146523341, b:157098276
TEST=Connect monitor support HDMI 2.0 display to RT2141's HDMI port.
     Confirm DUT did not output 4k@60fps that exceed HDMI 1.4 limitation.

Change-Id: Iab6b11b88d93e0b67d9cd8b0d42faa769122fc9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2352400
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Sean Paul <seanpaul@chromium.org>