UPSTREAM: media: vim2m: add buf_out_validate callback
Validate the field for an output buffer. This ensures that the
field is validated when the buffer is queued to a request, and
not when the request itself is queued, which is too late.

BUG=b:123551776
TEST=Build Kukui image.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
(cherry picked from commit ab7afaf33275acc2ccfaab3fe5b46870eaee9d55)
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

Change-Id: Iea89cbe9a8ccc48216880e50319eed66fa49a4e1
Reviewed-on: https://chromium-review.googlesource.com/1619663
Commit-Ready: Alexandre Courbot <acourbot@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>