UPSTREAM: media: vivid: add buf_out_validate callback
Validate the field for an output buffer. This ensures that the
field is validated when the buffer is queued to a request, and
not when the request itself is queued, which is too late.

BUG=b:123551776
TEST=Build Kukui image.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
(cherry picked from commit 1f2f510753becf607f9974a39a89ab1878f24b9d)
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

Change-Id: I0d070b457389411d6a3fcdb54301830f3e70c97c
Reviewed-on: https://chromium-review.googlesource.com/1619664
Commit-Ready: Alexandre Courbot <acourbot@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>