FROMLIST: interconnect: Revert to previous config if any request fails
When consumers report their bandwidth needs with icc_set_bw(), it's
possible that the requested amount of bandwidth is not available or just
the new configuration fails to apply on some path. In this case revert to
the previous configuration and propagate the error back to the consumers
to let them know that bandwidth is not available, hardware is busy or
whatever error is returned by the interconnect platform driver.

Signed-off-by: Georgi Djakov <georgi.djakov@linaro.org>
(am from https://lore.kernel.org/patchwork/patch/1032611/)
(also found at https://lkml.kernel.org/r/20190116161103.6937-9-georgi.djakov@linaro.org)

BUG=b:80591049
TEST=Boot cheza

Change-Id: Ia04ee69dd0f5c1c8b504908a981764033d1dd98e
Signed-off-by: Evan Green <evgreen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1418624
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>