BACKPORT: FROMGIT: media: vb2: Keep dma-buf buffers mapped until they are freed
When using vb2 for video decoding, dequeued capture buffers may still
be accessed by the hardware: this is the case when they are used as
reference frames for decoding subsequent frames.

When the buffer is imported with dma-buf, it needs to be mapped before
access. Until now, it was mapped when queuing and unmapped when
dequeuing, which doesn't work for access as a reference frames.

One way to solve this would be to map the buffer again when it is
needed as a reference, but the mapping/unmapping operations can
seriously impact performance. As a result, map the buffer once (when it
is first needed when queued) and keep it mapped until it is freed.

Reviewed-on: https://chromium-review.googlesource.com/334103
[Paul: Updated for mainline and changed commit message]

Signed-off-by: Pawel Osciak <posciak@chromium.org>
Signed-off-by: Paul Kocialkowski <paul.kocialkowski@bootlin.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
(cherry picked from commit 2cc1802f62e562611e86f04d9dae1337c824991e
 git://linuxtv.org/media_tree.git)

Commit d88937624351 ("media: videobuf2: remove unused variable")
removes the unused "i" variable left by this patch, but because of our
local change, commit 5b47324d523d ("CHROMIUM: vb2: Don't reuse USERPTR
buffers."), the variable is still needed, so I skipped it.

BUG=chromium:938781
TEST=Pre-CQ passes

Change-Id: I99723d66af8fe203124634fa010f5351218789bd
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1505082
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>