UPSTREAM: KVM: Expose the initial start value in grow_halt_poll_ns() as a module parameter
The hard-coded value 10000 in grow_halt_poll_ns() stands for the initial
start value when raising up vcpu->halt_poll_ns.
It actually sets the first timeout to the first polling session.
This value has significant effect on how tolerant we are to outliers.
On the standard case, higher value is better - we will spend more time
in the polling busyloop, handle events/interrupts faster and result
in better performance.
But on outliers it puts us in a busy loop that does nothing.
Even if the shrink factor is zero, we will still waste time on the first
iteration.
The optimal value changes between different workloads. It depends on
outliers rate and polling sessions length.
As this value has significant effect on the dynamic halt-polling
algorithm, it should be configurable and exposed.

Reviewed-by: Boris Ostrovsky <boris.ostrovsky@oracle.com>
Reviewed-by: Liran Alon <liran.alon@oracle.com>
Signed-off-by: Nir Weiner <nir.weiner@oracle.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 49113d360bdeb4dd916fb6bffbcc3e157422b6fd)

BUG=b:181105054
TEST=analyzed halt value distribution and per-VCPU polling metrics

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Change-Id: I21b64a09391b43590706014f0483f245a8422185
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3144440
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>