CHROMIUM: iwl7000: mac80211: Fix Tx aggregation session tear down with ITXQs
When mac80211 requests the low level driver to stop an ongoing
Tx aggregation, the low level driver is expected to call
ieee80211_stop_tx_ba_cb_irqsafe() to indicate that it is ready
to stop the session. The callback in turn schedules a worker
to complete the session tear down, which in turn also handles
the relevant state for the intermediate Tx queue.

However, as this flow in asynchronous, the intermediate queue
should be stopped and not continue servicing frames, as in
such a case frames that are dequeued would be marked as part
of an aggregation, although the aggregation is already been
stopped.

Fix this by stopping the intermediate Tx queue, before
calling the low level driver to stop the Tx aggregation.

Change-Id: Ia4d8eca3257dcd60a8f3c348ae1b5212dfc8588c
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: 61e70080aa86b127ac6e98579cdc43c1ce1c5463