FROMGIT: Bluetooth: hci_qca: Disable IBS state machine and flush Tx buffer
During hci down we observed IBS sleep commands are queued in the Tx
buffer and hci_uart_write_work is sending data to the chip which is
not required as the chip is powered off. This patch will disable IBS
and flush the Tx buffer before we turn off the chip.

Signed-off-by: Balakrishna Godavarthi <bgodavar@codeaurora.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit 035a960e7a279a59d74585105dcd263559d74f24
 git://git.kernel.org/pub/scm/linux/kernel/git/bluetooth/bluetooth-next.git master)

BUG=b:124761200
TEST=on cheza
  hciconfig hci0 down
  hciconfig hci0 up
  hciconfig | grep UP
    => UP RUNNING

Change-Id: Id51acd13e35c3850c22e978fe0fca36c73c77f11
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1358847
Reviewed-by: Sean Paul <seanpaul@chromium.org>