CHROMIUM: wilco_ec: Add event handling
The Wilco Embedded Controller can return extended events that
are not handled by standard ACPI objects. These events can
contain information about changes in EC controlled features,
such as the charger or battery getting plugged/unplugged.

These events are triggered with an ACPI Notify(0x90) and the
event data buffer is read through an ACPI method provided by
the BIOS which reads the event buffer from EC RAM.

These events are put into a queue which can be read by a userspace daemon
via a char device that implements read() and poll(). The char device
will appear at /dev/wilco_event{n}, where n is some integer.

This patch depends on a coreboot patch which adds the relevant
ACPI device, so this driver can pick it up:
https://review.coreboot.org/c/coreboot/+/31204

This is a fairly polished version that addressed all previous concerns.
Please look it over, and if we like it then I can send it upstream.

BUG=b:119046283
TEST=manual: Run the simple python script from
https://gist.github.com/8a55050e353b7027ca3009074eec4200
to poll() and read() the node. You can generate events
in many ways, the easiest are to press fn+Esc or to
plug/unplug the battery (with the AC in).

Change-Id: Ie63c52a1c38bbc60668fad3ebf9bd0ed529580b8
Signed-off-by: Nick Crews <ncrews@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1356083
Commit-Ready: Duncan Laurie <dlaurie@google.com>