FROMLIST: wilco_ec: Add support for raw commands in debugfs
(am from https://lkml.org/lkml/2019/2/7/1016)

Add a debugfs attribute that allows sending raw commands to the EC.
This is useful for development and debug but should not be enabled
in a production environment.

BUG=b:119046283
TEST=manual:
Get the EC firmware build date
First send the request command
> echo 00 f0 38 00 03 00 > raw
Then read the result. "12/21/18" is in the middle of the response
> cat raw
00 31 32 2f 32 31 2f 31 38 00 00 0f 01 00 01 00  .12/21/18.......

Change-Id: I3fe514de07c9b50a5dcd388bb5f0d9c8047a136e
Signed-off-by: Nick Crews <ncrews@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1323832
Commit-Ready: Duncan Laurie <dlaurie@google.com>
Tested-by: Nick Crews <ncrews@chromium.org>
Reviewed-by: Nick Crews <ncrews@chromium.org>