BACKPORT: platform/chrome: Add new driver for Wilco EC
This EC is an incompatible variant of the typical Chrome OS embedded
controller.  It uses the same low-level communication and a similar
protocol with some significant differences.  The EC firmware does
not support the same mailbox commands so it is not registered as a
cros_ec device type.  This commit exports the wilco_ec_mailbox()
function so that other modules can use it to communicate with the EC.

Signed-off-by: Duncan Laurie <dlaurie@google.com>
Signed-off-by: Nick Crews <ncrews@chromium.org>
[Fix the sparse warning: symbol 'wilco_ec_transfer' was not declared]
Signed-off-by: Wei Yongjun <weiyongjun1@huawei.com>
[Fix Kconfig dependencies for wilco_ec]
Reported-by: Randy Dunlap <rdunlap@infradead.org>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>

(cherry picked from commit 7b3d4f44abf0e7a1ba762c8a9c99a8b39ee0c8b1)

[ncrews: resolve delta between the cherry-picked UPSTREAM commit and
 the current Chromium state caused by the following: patches:
 15b95780aadc FROMLIST: wilco_ec: Add new driver for Wilco EC
 c0a3d5796a00 CHROMIUM: wilco_ec: Fix check of NO_RESPONSE flag in mailbox()
 cf8a809363c0 FROMLIST: wilco_ec: Fix Kconfig dependencies
 f34acfb57c99 FROMLIST: wilco_ec: Fix wilco_ec core driver.]

BUG=b:119046283
TEST=The kernel still builds and works.

Change-Id: I149a0f11786b4df70a13534fca7b387e23be93ec
Signed-off-by: Nick Crews <ncrews@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1551944
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>