BACKPORT: platform/chrome: wilco_ec: Add support for raw commands in debugfs
Add a debugfs attribute that allows sending raw commands to the EC.
This is useful for development and debug but should not be enabled
in a production environment.

To test:
Get the EC firmware build date
First send the request command
> echo 00 f0 38 00 03 00 > raw
Then read the result. "12/21/18" is in the middle of the response
> cat raw
00 31 32 2f 32 31 2f 31 38 00 00 0f 01 00 01 00  .12/21/18.......

Get the EC firmware build date
First send the request command
> echo 00 f0 38 00 03 00 > raw
Then read the result. "12/21/18" is in the middle of the response
> cat raw
00 31 32 2f 32 31 2f 31 38 00 00 0f 01 00 01 00  .12/21/18.......

Signed-off-by: Duncan Laurie <dlaurie@google.com>
Signed-off-by: Nick Crews <ncrews@chromium.org>
[Fix off-by-one error in wilco_ec/debugfs.c]
Reported-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>

(cherry picked from commit b787bb126cbcd73754bcbc055ae9f804ac576e4a)

[ncrews: resolve delta between the cherry-picked UPSTREAM commit and
 the current Chromium state caused by the following patches:
 117929663c FROMLIST: wilco_ec: Add support for raw commands in debugfs]

BUG=b:119046283
TEST=The kernel still builds and works.

Change-Id: If401b25fc802711fd1d680cb396e4a87cdfcc8fd
Signed-off-by: Nick Crews <ncrews@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1553114
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>