BACKPORT: platform/chrome: wilco_ec: Add RTC driver
This Embedded Controller has an internal RTC that is exposed
as a standard RTC class driver with read/write functionality.

The driver is added to the drivers/rtc/ so that the maintainer of that
directory will be able to comment on this change, as that maintainer is
the expert on this system. In addition, the driver code is called
indirectly after a corresponding device is registered from core.c,
as opposed to core.c registering the driver callbacks directly.

To test:
> hwclock --show --rtc /dev/rtc1
2007-12-31 16:01:20.460959-08:00
> hwclock --systohc --rtc /dev/rtc1
> hwclock --show --rtc /dev/rtc1
2018-11-29 17:08:00.780793-08:00

> hwclock --show --rtc /dev/rtc1
2007-12-31 16:01:20.460959-08:00
> hwclock --systohc --rtc /dev/rtc1
> hwclock --show --rtc /dev/rtc1
2018-11-29 17:08:00.780793-08:00

Signed-off-by: Duncan Laurie <dlaurie@google.com>
Signed-off-by: Nick Crews <ncrews@chromium.org>
Acked-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
[Fix the sparse warning: symbol 'wilco_ec_rtc_read/write' was not declared]
Signed-off-by: Wei Yongjun <weiyongjun1@huawei.com>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>

(cherry picked from commit 0d2f2a3da1f2a9ebeb66bb03073dd149fccf1bdd)

[ncrews: resolve delta between the cherry-picked UPSTREAM commit and
 the current Chromium state caused by the following: patches:
 6448fd21910e FROMLIST: wilco_ec: Add RTC driver]

BUG=b:119046283
TEST=The kernel still builds and works.

Change-Id: I204dae74e9d8e3abaad9c81e3462b20b74807f2d
Signed-off-by: Nick Crews <ncrews@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1553118
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>