FROMLIST: wilco_ec: Fix Kconfig dependencies
In the initial version of the Wilco EC Driver, the
dependency order was wrong. It before was possible to
select CONFIG_WILCO_EC and CONFIG_CROS_EC_LPC without
having CONFIG_CROS_EC_LPC_MEC. This was wrong, since
WILCO_EC depends upon CONFIG CROS_EC_LPC_MEC, not the
other way around.

I'm working on getting this into the driver that is
about to be merged in the 5.1 kernel, hence the
FROMLIST: but I wanted to get this on here too for visibility.

BUG=b:1190462
TEST=manual

Change-Id: I1758082af096bfc16a462e5e15cafa57612027eb
Signed-off-by: Nick Crews <ncrews@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1469042
Reviewed-by: Guenter Roeck <groeck@chromium.org>