FROMGIT: PM / Domains: Mark "name" const in dev_pm_domain_attach_by_name()
As of the patch ("PM / Domains: Mark "name" const in
genpd_dev_pm_attach_by_name()") it's clear that the name in
dev_pm_domain_attach_by_name() can be const.  Mark it as so.  This
allows drivers to pass in a name that was declared "const" in a
driver.

Fixes: 27dceb81f445 ("PM / Domains: Introduce dev_pm_domain_attach_by_name()")
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit eeb35df05244c268cd69b425edf6dc6a49ee7ab4
 git://git.kernel.org/pub/scm/linux/kernel/git/rafael/linux-pm.git linux-next)

BUG=b:80592702
TEST=Device doesn't crash

Change-Id: I30685a385a6c12fb1f9848f039a41205455ba340
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1500117
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>